+++ CHANGELOG.md
@@ -11,10 +11,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Added a simple RC model for HVAC time series (`#12`, `!6`)
+- Added simpler functionality for dependent methods (`#18`, `!12`)
+- Basic documentation of the package (`#13`, `!8`)
+
+### Changed
+- Restructured entire architecture towards a pipeline- and strategy-based approach to make methods more flexible (`#18`, `!12`)
 
 ## [0.1.0] Initial Release - Hello ENS-Repo-Template - 2024-11-04
 ### Added
-- Initial setup for project.
+- Initial setup of the project
 
 ---